requirements.txt
amqp==2.5.1
appdirs==1.4.3
architect==0.5.6
asgiref==3.5.2
attrs==19.1.0
backports.csv==1.0.7
backports.zoneinfo==0.2.1
billiard==3.6.1.0
cached-property==1.5.1
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-admin-honeypot==1.1.0
django-chunked-iterator==0.6.1
django-compat==1.0.15
django-crispy-forms==1.7.2
django-datatables-view==1.19.1
django-import-export==1.2.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-session-timeout==0.1.0
djangorestframework==3.10.3
drf-yasg==1.17.1
et-xmlfile==1.0.1
fs==2.4.10
html5lib==1.0.1
idna==2.8
importlib-metadata==0.22
inflection==0.5.1
install==1.3.5
isodate==0.6.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.1.1
kombu==4.6.4
lxml==4.4.1
MarkupSafe==2.1.1
more-itertools==7.2.0
mysqlclient==2.1.0
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.3
packaging==21.3
Pillow==7.1.2
pyparsing==3.0.7
PyPDF2==1.26.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
reportlab==3.6.6
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.5.12
simplejson==3.17.6
six==1.12.0
sqlparse==0.3.0
suds-py3==1.3.3.0
tablib==0.13.0
uritemplate==4.1.1
urllib3==1.25.3
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
zeep==3.4.0
zipp==0.6.0
